fix login action
